+++ Project2_backend/src/main/java/com/revature/Project2_backend/controllers/UserResource.java
@@ -46,9 +46,13 @@ public class UserResource {
 
     @Transactional
     @DeleteMapping("/userDelete/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable("id") Long userId){
+    public ResponseEntity<?> deleteUser(@PathVariable("id") Long userId) {
+      User findUser = userService.findUserById(userId);
+      if(findUser != null){
         userService.deleteUser(userId);
         return new ResponseEntity<>(HttpStatus.OK);
+      }
+      return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping("/login")
